Finish LL Spi proxy

+++ LLProxys/spi/SpiProxy.h
@@ -20,9 +20,9 @@ typedef struct
 
     void (*configure)(void *self, void *spi_handle, void *csPort, uint16_t csPin);
 
-    void (*transmit)(void *self, uint8_t *data, uint16_t size);
-    void (*receive)(void *self, uint8_t *data, uint16_t size);
-    void (*transmit_receive)(void *self, uint8_t *txData, uint8_t *rxData, uint16_t size);
+    void (*transmit)(void *self, uint8_t *data, size_t size);
+    void (*receive)(void *self, uint8_t *data, size_t size);
+    void (*transmit_receive)(void *self, uint8_t *txData, uint8_t *rxData, size_t size);
 
 } Spi;
